+++ NEWS.md
@@ -1,6 +1,10 @@
 Red Panda C++ Version 1.1.3
 
   - fix: wrong auto indent calculation for comments
+  - enhancement: position caret at end of the line of folded code block
+  - enhancement: copy the whole folded code block
+  - enhancement: delete the whole folded code block
+  - fix: correctly update the folding state of code block, when deleted
 
 Red Panda C++ Version 1.1.2
   - enhancement: use different color to differenciate folder and headers in completion popup window